ai-ensemble/requirements.txt
fastapi
uvicorn
python-multipart
requests
ultralytics
transformers
torch
Pillow
numpy
scipy
